restrict gpu visibility

+++ example4.py
@@ -30,7 +30,7 @@ def get_encodes(x):
 
 
 data_node = (tf.data.TextLineDataset(train_fp).batch(batch_size)
-             .map(lambda x: tf.py_func(get_encodes, [x], [tf.float32, tf.int32], name='train_mktokens_fn'),
+             .map(lambda x: tf.py_func(get_encodes, [x], [tf.float32, tf.int64], name='train_mktokens_fn'),
                   num_parallel_calls=num_parallel_calls)
              .make_one_shot_iterator().get_next())
